requirements.txt
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
numpy==1.26.4
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core
gdown
librosa
inflect

black
isort
pre-commit
